[FIX] account_move_tier_validation: fix _get_to_validate_message_name
