+++ account_move_tier_validation/models/__init__.py
@@ -1,3 +1,2 @@
 from . import account_move
 from . import tier_definition
-from . import tier_validation
