+++ account_move_tier_validation/models/account_move.py
@@ -1,7 +1,7 @@
 # Copyright <2020> PESOL <info@pesol.es>
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl)
 
-from odoo import models
+from odoo import _, models
 
 
 class AccountMove(models.Model):
@@ -9,3 +9,15 @@ class AccountMove(models.Model):
     _inherit = ["account.move", "tier.validation"]
     _state_from = ["draft"]
     _state_to = ["posted"]
+
+    def _get_to_validate_message_name(self):
+        name = super(AccountMove, self)._get_to_validate_message_name()
+        if self.type == "in_invoice":
+            name = _("Bill")
+        elif self.type == "in_refund":
+            name = _("Refund")
+        elif self.type == "out_invoice":
+            name = _("Invoice")
+        elif self.type == "out_refund":
+            name = _("Credit Note")
+        return name
